Swapped tilesz and tilesz2.

+++ src/haven/MapMesh.java
@@ -120,7 +120,7 @@ public class MapMesh implements Rendered, Disposable {
 	public MapSurface() {
 	    for(int y = vs.ul.y; y < vs.br.y; y++) {
 		for(int x = vs.ul.x; x < vs.br.x; x++) {
-		    surf[vs.o(x, y)] = new Vertex(x * tilesz.x, y * -tilesz.y, map.getz(ul.add(x, y)));
+		    surf[vs.o(x, y)] = new Vertex(x * (float)tilesz.x, y * -(float)tilesz.y, map.getz(ul.add(x, y)));
 		}
 	    }
 	    for(int y = ts.ul.y; y < ts.br.y; y++) {
@@ -350,8 +350,8 @@ public class MapMesh implements Rendered, Disposable {
 	    {
 		Coord tult = null, tbrt = null;
 		for(Coord3f corn : new Coord3f[] {ul, new Coord3f(ul.x, br.y, 0), br, new Coord3f(br.x, ul.y, 0)}) {
-		    float cx = (float)(cc.x + co * corn.x - si * corn.y) / tilesz.x;
-		    float cy = (float)(cc.y + co * corn.y + si * corn.x) / tilesz.y;
+		    float cx = (float)((cc.x + co * corn.x - si * corn.y) / tilesz.x);
+		    float cy = (float)((cc.y + co * corn.y + si * corn.x) / tilesz.y);
 		    if(tult == null) {
 			tult = new Coord((int)Math.floor(cx), (int)Math.floor(cy));
 			tbrt = new Coord((int)Math.ceil(cx), (int)Math.ceil(cy));
@@ -372,8 +372,8 @@ public class MapMesh implements Rendered, Disposable {
 		    public void faces(MapMesh m, Tiler.MPart d) {
 			Coord3f[] texc = new Coord3f[d.v.length];
 			for(int i = 0; i < d.v.length; i++) {
-			    texc[i] = new Coord3f(((m.ul.x + d.lc.x + d.tcx[i]) * tilesz.x) - (float)cc.x,
-						  ((m.ul.y + d.lc.y + d.tcy[i]) * tilesz.y) - (float)cc.y,
+			    texc[i] = new Coord3f((float)(((m.ul.x + d.lc.x + d.tcx[i]) * tilesz.x) - cc.x),
+						  (float)(((m.ul.y + d.lc.y + d.tcy[i]) * tilesz.y) - cc.y),
 						  0);
 			    texc[i] = new Coord3f(co * texc[i].x + si * texc[i].y,
 						  co * texc[i].y - si * texc[i].x,
@@ -412,7 +412,7 @@ public class MapMesh implements Rendered, Disposable {
 				continue;
 			    if((mv[i] = cv.get(d.v[i])) == null) {
 				cv.put(d.v[i], mv[i] = new MeshVertex(buf, d.v[i]));
-				mv[i].pos = mv[i].pos.add((m.ul.x * tilesz.x) - (float)cc.x, (float)cc.y - (m.ul.y * tilesz.y), -cz);
+				mv[i].pos = mv[i].pos.add((float)((m.ul.x * tilesz.x) - cc.x), (float)(cc.y - (m.ul.y * tilesz.y)), -cz);
 				ta.set(mv[i], texc[i]);
 			    }
 			}
